# compare-commits.sh: support both llama-bench and test-backend-ops