@yeahdongcn 2025-07-05
Hi @JohannesGaessler @slaren @ggerganov I’ve merged https://github.com/ggml-org/llama.cpp/pull/14368 into master. Could you please continue reviewing this one when you have a moment? Thanks!